Wrote proposal map tab active badge functionality (#60); did a little reorganization.
