+++ static-src/components/App.js
@@ -3,7 +3,7 @@
 */
 
 import React, { useState, useRef, useEffect } from 'react';
-import { useDispatch, useSelector } from 'react-redux';
+import { useDispatch } from 'react-redux';
 import { BrowserRouter, Route, Redirect } from 'react-router-dom';
 import styled from 'styled-components';
 
@@ -13,14 +13,7 @@ import Header from './Header';
 import LeafletMap from './LeafletMap';
 import VisibleMapArea from './VisibleMapArea';
 import Panel from './Panel';
-import PanelView from './PanelView';
-import HTMLContent from './HTMLContent';
-import ProposalMapsPanelViewHeader from './ProposalMapsPanelViewHeader';
-import ProposalMapsPanelViewBody from './ProposalMapsPanelViewBody';
-import NarrativesPanelViewHeader from './NarrativesPanelViewHeader';
-import NarrativesPanelViewBody from './NarrativesPanelViewBody';
-import Modal from './Modal';
-import FeedbackForm from './FeedbackForm';
+import Modals from './Modals';
 
 
 /*
@@ -34,12 +27,6 @@ export default function App() {
   const visibleMapAreaElementRef = useRef();
   const leafletMapRef = useRef();
   const dispatch = useDispatch();
-  const {
-    introduction,
-    bibliography,
-    credits,
-    feedback,
-  } = useSelector((state) => state.contentAreaContent);
 
 
   //  Downloads initial content and sets initial state
@@ -57,34 +44,6 @@ export default function App() {
     initApp();
   }, [dispatch]);
 
-  //  Modals
-  const modals = [
-
-    <Modal
-      path="/bibliography"
-      title="Bibliography"
-      content={<HTMLContent content={bibliography} />}
-    />,
-
-    <Modal
-      path="/credits"
-      title="Credits"
-      content={<HTMLContent content={credits} />}
-    />,
-
-    <Modal
-      path="/feedback"
-      title="Feedback"
-      content={(
-        <>
-          <HTMLContent content={feedback} />
-          <FeedbackForm />
-        </>
-      )}
-    />,
-
-  ];
-
   return (
     <LeafletMapContext.Provider value={leafletMapRef}>
       <StyledApp>
@@ -101,37 +60,9 @@ export default function App() {
                   visibleMapAreaElementRef={visibleMapAreaElementRef}
                   ref={leafletMapRef}
                 />
-                <Panel>
-
-                  <PanelView
-                    path="/introduction"
-                    title="Introduction"
-                    bodyContent={<HTMLContent content={introduction} />}
-                  />
-
-                  <PanelView
-                    path="/proposal-maps"
-                    title="Proposal Maps"
-                    headerContent={<ProposalMapsPanelViewHeader />}
-                    bodyContent={<ProposalMapsPanelViewBody />}
-                  />
-
-                  <PanelView
-                    path="/narratives"
-                    title="Narratives"
-                    headerContent={<NarrativesPanelViewHeader pathPrefix="/narratives/" />}
-                    bodyContent={<NarrativesPanelViewBody />}
-                  />
-
-                </Panel>
+                <Panel />
                 <VisibleMapArea ref={visibleMapAreaElementRef} />
-                { modals.map((modal) => (
-                  <Route
-                    path={modal.props.path}
-                    key={modal.props.path}
-                    render={() => modal}
-                  />
-                )) }
+                <Modals />
               </StyledBody>
               <Route
                 path="/"
